# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Compliance OS — GRC tool API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entities/{code}:
    get:
      operationId: getEntity
      tags: [entities]
      summary: Get a single entity
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
        "404":
          description: Not found

  # ─── FRAMEWORKS ─────────────────────────────────────────────────────────────
  /frameworks:
    get:
      operationId: listFrameworks
      tags: [frameworks]
      summary: List compliance frameworks
      responses:
        "200":
          description: Frameworks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Framework"
    post:
      operationId: createFramework
      tags: [frameworks]
      summary: Create a framework
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrameworkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Framework"
  /frameworks/{id}:
    get:
      operationId: getFramework
      tags: [frameworks]
      summary: Get a framework
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Framework
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Framework"
        "404":
          description: Not found
    patch:
      operationId: updateFramework
      tags: [frameworks]
      summary: Update a framework
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrameworkUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Framework"

  # ─── CONTROLS ───────────────────────────────────────────────────────────────
...
  /evidence:
    get:
      operationId: listEvidence
      tags: [evidence]
      summary: List evidence requests
      parameters:
        - name: entityCode
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: priority
          in: query
          schema:
            type: ["string", "null"]
        - name: frameworkId
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["number", "null"]
        - name: offset
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: Evidence requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvidenceRequest"
    post:
      operationId: createEvidence
      tags: [evidence]
      summary: Create evidence request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceRequest"
  /evidence/{id}:
    get:
      operationId: getEvidence
      tags: [evidence]
      summary: Get an evidence request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Evidence request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceRequest"
        "404":
          description: Not found
    patch:
      operationId: updateEvidence
      tags: [evidence]
      summary: Update evidence request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceRequest"
    delete:
      operationId: deleteEvidence
      tags: [evidence]
      summary: Delete evidence request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── POLICIES ───────────────────────────────────────────────────────────────
  /policies:
    get:
      operationId: listPolicies
      tags: [policies]
      summary: List policies
      parameters:
        - name: entityCode
          in: query
          schema:
            type: ["string", "null"]
        - name: frameworkId
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
    post:
      operationId: createPolicy
      tags: [policies]
      summary: Create a policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /policies/{id}:
    get:
      operationId: getPolicy
      tags: [policies]
      summary: Get a policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "404":
          description: Not found
    patch:
      operationId: updatePolicy
      tags: [policies]
      summary: Update a policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    delete:
      operationId: deletePolicy
      tags: [policies]
      summary: Delete a policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── AOCs ───────────────────────────────────────────────────────────────────
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List third-party vendors / TPSPs
      parameters:
        - name: entityCode
          in: query
          schema:
            type: ["string", "null"]
        - name: riskLevel
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── ACTIVITY ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Entity:
      type: object
      required: [code, name, legalName, merchantLevel, saqType, cdeScope, isActive, createdAt]
      properties:
        code:
          type: string
        name:
          type: string
        legalName:
          type: string
        merchantLevel:
          type: ["string", "null"]
        saqType:
          type: ["string", "null"]
        cdeScope:
          type: ["string", "null"]
        lastAocDate:
          type: ["string", "null"]
        nextAocDate:
          type: ["string", "null"]
        qsaCompany:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string

    # ─── Framework ──────────────────────────────────────────────────────────
    Framework:
      type: object
      required: [id, code, name, version, priority, status, summary, domainsCount, nextMilestone, createdAt]
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        version:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [active, monitored, retired]
        summary:
          type: string
        domainsCount:
          type: number
        nextMilestone:
          type: ["string", "null"]
        owner:
          type: ["string", "null"]
        createdAt:
          type: string
    FrameworkInput:
      type: object
      required: [code, name, version, priority, status, summary, domainsCount]
      properties:
        code:
          type: string
        name:
          type: string
        version:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [active, monitored, retired]
        summary:
          type: string
        domainsCount:
          type: number
        nextMilestone:
          type: string
        owner:
          type: string
    FrameworkUpdate:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [active, monitored, retired]
        summary:
          type: string
        domainsCount:
          type: number
        nextMilestone:
          type: string
        owner:
          type: string

    # ─── Control ────────────────────────────────────────────────────────────
...
    EvidenceRequest:
      type: object
      required: [id, code, entityCode, controlRef, frameworkCode, title, status, priority, requestedAt, createdAt]
      properties:
        id:
          type: string
        code:
          type: string
        assessmentId:
          type: ["string", "null"]
        controlId:
          type: ["string", "null"]
        controlRef:
          type: string
        controlName:
          type: ["string", "null"]
        frameworkCode:
          type: string
        frameworkName:
          type: ["string", "null"]
        entityCode:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [requested, in-progress, submitted, approved, rejected]
        priority:
          type: string
          enum: [critical, high, medium, low]
        assignee:
          type: ["string", "null"]
        requestedBy:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        requestedAt:
          type: string
        approvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    EvidenceInput:
      type: object
      required: [entityCode, controlRef, frameworkCode, title, priority]
      properties:
        entityCode:
          type: string
        controlId:
          type: string
        controlRef:
          type: string
        frameworkCode:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]
        assignee:
          type: string
        requestedBy:
          type: string
        dueDate:
          type: string
        assessmentId:
          type: string
    EvidenceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [requested, in-progress, submitted, approved, rejected]
        priority:
          type: string
          enum: [critical, high, medium, low]
        assignee:
          type: string
        dueDate:
          type: string
        description:
          type: string
        title:
          type: string

    # ─── Policy ─────────────────────────────────────────────────────────────
    Policy:
      type: object
      required: [id, code, name, owner, version, status, createdAt]
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        owner:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [draft, current, review-due, overdue, retired]
        effectiveDate:
          type: ["string", "null"]
        reviewDate:
          type: ["string", "null"]
        pages:
          type: ["number", "null"]
        frameworks:
          type: array
          items:
            type: string
        entities:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    PolicyInput:
      type: object
      required: [code, name, owner, version, status]
      properties:
        code:
          type: string
        name:
          type: string
        owner:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [draft, current, review-due, overdue, retired]
        effectiveDate:
          type: string
        reviewDate:
          type: string
        pages:
          type: number
        frameworks:
          type: array
          items:
            type: string
        entities:
          type: array
          items:
            type: string
        description:
          type: string
    PolicyUpdate:
      type: object
      properties:
        name:
          type: string
        owner:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [draft, current, review-due, overdue, retired]
        effectiveDate:
          type: string
        reviewDate:
          type: string
        pages:
          type: number
        frameworks:
          type: array
          items:
            type: string
        entities:
          type: array
          items:
            type: string
        description:
          type: string

    # ─── AOC ────────────────────────────────────────────────────────────────
...
    Vendor:
      type: object
      required: [id, name, serviceType, riskLevel, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        serviceType:
          type: string
        description:
          type: ["string", "null"]
        riskLevel:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [active, inactive, under-review]
        entities:
          type: array
          items:
            type: string
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        lastReviewDate:
          type: ["string", "null"]
        nextReviewDate:
          type: ["string", "null"]
        hasPciCertification:
          type: boolean
        certificationExpiry:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    VendorInput:
      type: object
      required: [name, serviceType, riskLevel]
      properties:
        name:
          type: string
        serviceType:
          type: string
        description:
          type: string
        riskLevel:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [active, inactive, under-review]
        entities:
          type: array
          items:
            type: string
        contactName:
          type: string
        contactEmail:
          type: string
        lastReviewDate:
          type: string
        nextReviewDate:
          type: string
        hasPciCertification:
          type: boolean
        certificationExpiry:
          type: string
        notes:
          type: string
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        serviceType:
          type: string
        description:
          type: string
        riskLevel:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [active, inactive, under-review]
        entities:
          type: array
          items:
            type: string
        contactName:
          type: string
        contactEmail:
          type: string
        lastReviewDate:
          type: string
        nextReviewDate:
          type: string
        hasPciCertification:
          type: boolean
        certificationExpiry:
          type: string
        notes:
          type: string

    # ─── Activity ───────────────────────────────────────────────────────────
...